backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic-settings==2.2.0
celery==5.3.6
redis==5.0.3
httpx==0.27.0
twilio==9.0.4
openai==1.12.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
email-validator==2.1.0
bcrypt==4.0.1
passlib==1.7.4
python-socketio[asyncio]==5.11.4